% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cf_resources}
\alias{cf_resources}
\title{Community and Family Resources}
\format{
A data frame with 10 rows and 8 variables:
\describe{
  \item{address}{Street address for the facility}
  \item{city}{Character string containing city name}
  \item{state}{State Abbreviation}
  \item{zip}{5 digit zip code}
  \item{name}{Facility Name, only 3 locations have unique names}
  \item{search_address}{address used in the Google API geocoding}
  \item{longitude}{geographic Longitude}
  \item{latitude}{geographic Latitude}
  \item{classification}{classification column}
}
}
\source{
\url{http://www.cfrhelps.org/our-locations}
}
\usage{
cf_resources
}
\description{
Dataset was scraped from the Community and Family Resources website.
This database contains Community and Family Resources location data for the State of Iowa.
}
\examples{
library(ggplot2)
library(dplyr)
cf_resources \%>\%
 ggplot() +
 geom_point(aes(x = longitude, y = latitude))

library(leaflet)
library(sf)
cf_resources \%>\%
 leaflet() \%>\%
 addTiles() \%>\%
 addPolygons(data = ia_counties,
             weight = 1, color="#333333") \%>\%
 addCircleMarkers(lng = ~longitude, lat = ~latitude,
                  radius = 1, stroke = 0.1)
}
\keyword{datasets}
