% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{churches}
\alias{churches}
\title{Location of churches in Iowa}
\format{A data frame with 5469 rows and 9 variables:
\describe{
   \item{ID}{identifier, not quite the row number XXX we should delete this column and move the GNIS_ID up.}
   \item{NAME}{name of the church}
   \item{CATEGORY}{value of `Iowa physical, cultural and historic features`}
   \item{TYPE}{type of location, constant value of `Cultural`.}
   \item{CLASS}{type of structure, constant value of `CHURCH`.}
   \item{DESCRIPTION}{closer description of the location, if available}
   \item{COUNTY}{name of the county}
   \item{GNIS_ID}{USGS identifier}
  \item{geometry}{sf point object of geographic locations of churches in Iowa.}
}}
\source{
\url{https://iowa.hometownlocator.com/features/cultural,class,church.cfm}
}
\usage{
churches
}
\description{
Dataset was scraped by Masoud Nosrati from the IA Hometown Locator in Mar 2020.
}
\examples{
# Map of churches in Story county using ggplot2
library(ggplot2)
library(dplyr) # for the pipe

churches \%>\%
 # filter(COUNTY == "Story county) \%>\%
  ggplot() +
    geom_sf(data = ia_counties) +
    geom_sf() +
    ggthemes::theme_map()

# leaflet map
library(leaflet)
library(sf)

churches \%>\%
  leaflet() \%>\%
    addTiles() \%>\%
    setView(-93.6498803, 42.0275751, zoom = 8) \%>\%
    addPolygons(data = st_transform(ia_counties, crs='+proj=longlat +datum=WGS84'),
                weight = 1, color="#333333") \%>\%
    addCircleMarkers(radius = 1, stroke = 0.1,
                     label = ~NAME)
}
\keyword{datasets}
