% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colleges}
\alias{colleges}
\title{Location of colleges and universities in Iowa}
\format{A data frame with 57 rows and 11 variables:
\describe{
   \item{SCHOOL}{name of school}
   \item{CITY}{name of the clinic}
   \item{CONTROL}{type of institution by funding}
   \item{TYPE}{type of institution by degree}
   \item{ENROLMENT}{number of students enrolled (in Spring 2012)}
   \item{FOUNDED}{year in which the institution was founded}
   \item{address}{address used in search string for Google geocoding}
   \item{Latitude}{geographic latitude}
   \item{Longitude}{geographic longitude}
}}
\source{
\url{https://en.wikipedia.org/wiki/List_of_colleges_and_universities_in_Iowa}
}
\usage{
colleges
}
\description{
Dataset was scraped by Masoud Nosrati from wikipedia in Mar 2020, geocoding by Heike Hofmann through Google API.
}
\examples{
# Map of rural health clinics in Iowa using ggplot2
library(ggplot2)
library(dplyr) # for the pipe

colleges \%>\%
  ggplot() +
    geom_point(aes(x = Longitude, y = Latitude))

# leaflet map
library(leaflet)
library(sf)

colleges \%>\%
  leaflet() \%>\%
    addTiles() \%>\%
    addPolygons(data = st_transform(ia_counties, crs='+proj=longlat +datum=WGS84'),
                weight = 1, color="#333333") \%>\%
    addCircleMarkers(lng = ~Longitude, lat = ~Latitude,
                     radius = 1, stroke = 0.1,
                     label = ~SCHOOL)
}
\keyword{datasets}
