% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cross_mental_health}
\alias{cross_mental_health}
\title{Cross Mental Health Data}
\format{A data frame with 36 rows and 5 variables:
\describe{
  \item{name}{Name of the accesss point}
  \item{address}{Address of the building}
  \item{state}{state abbreviation}
  \item{zip}{zip code}
  \item{phone}{Phone number of the building}
  \item{longitude}{geographic Longitude}
  \item{latitude}{geographic Latitude}
  \item{classification}{column used for classifying the data set}
}}
\source{
\url{https://crossmentalhealth.org/wp-content/uploads/2020/03/Access-Points.pdf}
}
\usage{
cross_mental_health
}
\description{
CROSS has designated access points for adult mental health and disability services that are listed here
}
\examples{
library(dplyr)

# Leaflet map of Cross Mental Health buildings in Iowa
library(leaflet)
library(sf)

cross_mental_health \%>\%
 leaflet() \%>\%
 addTiles() \%>\%
 addPolygons(data = ia_counties,
             weight = 1, color="#333333") \%>\%
 addCircleMarkers(radius = 1, stroke = 0.1)
}
\keyword{datasets}
