% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gambling}
\alias{gambling}
\title{Licensed Substance Use Disorder/Problem Gambling Programs}
\format{
A data frame with 105 rows and 9 variables:
\describe{
  \item{name}{Name of the accesss point}
  \item{person}{Address of the building}
  \item{email}{email id}
  \item{address}{Full address}
  \item{phone}{Phone number}
  \item{dataset}{Dataset name}
  \item{classification}{domain}
  \item{longitude}{geographic Latitude}
  \item{latitude}{geographic Longitude}
}
}
\source{
\url{https://idph.iowa.gov/Portals/1/userfiles/166/Licensure/All\%20Licensed\%20Substance\%20Use\%20Disorder\%20-\%20Problem\%20Gambling\%20Program\%27s\%20List.pdf}
}
\usage{
gambling
}
\description{
List of  Licensed Substance Use Disorder/Problem Gambling Programs
}
\keyword{datasets}
