% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meetings.R
\name{get_meetings}
\alias{get_meetings}
\title{Get set of meetings for a time range based on schedule}
\usage{
get_meetings(
  from = now(),
  to = now() + days(1),
  type = c("Alcoholics Anonymous", "Narcotics Anonymous")
)
}
\arguments{
\item{from}{earliest date (and time, if specified)}

\item{to}{latest date}

\item{type}{type of meeting see `unique(meetings$type)`}
}
\value{
data frame
}
\description{
The `meetings` object consist of a schedule of AA meetings and locations
in Iowa. This function creates a data frame of all available meetings and locations
for a specified time range.
}
\examples{
library(leaflet)
library(sf)

# AA meetings in Iowa for the next 24h from
lubridate::now()
# color locations by day
pal <- colorFactor(
         palette = RColorBrewer::brewer.pal(n=7, name="Paired"),
         domain = levels(meetings$day)
       )
get_meetings() \%>\%
  leaflet() \%>\%
  addTiles() \%>\%
  addPolygons(data = st_transform(ia_counties, crs='+proj=longlat +datum=WGS84'),
              weight = 1, color="#888888") \%>\%
  addCircleMarkers(lng = ~longitude, lat = ~latitude,
                   color=~pal(day), label=~meeting,
                   radius = 1, stroke = 0.1) \%>\%
  addLegend(pal = pal, values = levels(meetings$day))
}
