% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{health.clinics}
\alias{health.clinics}
\title{Location of rural health clinics in Iowa}
\format{A data frame with 146 rows and 11 variables:
\describe{
   \item{name}{name of the clinic}
   \item{address}{formatted addresses - some are duplicates - XXX look into}
   \item{state}{state abbreviation}
   \item{county}{name of the county}
   \item{zip}{zip code}
   \item{country}{country where facility is located}
   \item{result_num}{146 0s XXX delete column?}
   \item{status}{146 OKs XXX delete column?}
   \item{longitude}{geographic longitude}
   \item{latitude}{geographic latitude}
   \item{classification}{column used for classifying the data set}
}}
\source{
\url{https://iarhc.org/find-a-rural-health-clinic?view=map}
}
\usage{
health.clinics
}
\description{
Dataset was scraped by Masoud Nosrati from the Iowa Association of Rural Health Clinics in Mar 2020, geocoding by Andrew Maloney through QGIS.
}
\examples{
# Map of rural health clinics in Iowa using ggplot2
library(ggplot2)
library(dplyr) # for the pipe

health.clinics \%>\%
  ggplot() +
    geom_point(aes(x = Longitude, y = Latitude))

# leaflet map
library(leaflet)
library(sf)

health.clinics \%>\%
  leaflet() \%>\%
    addTiles() \%>\%
    addPolygons(data = ia_counties,
                weight = 1, color="#333333") \%>\%
    addCircleMarkers(lng = ~Longitude, lat = ~Latitude,
                     radius = 1, stroke = 0.1,
                     label = ~NAME)
}
\keyword{datasets}
