% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hospital_buildings}
\alias{hospital_buildings}
\title{Iowa Hospital Buildings}
\format{A data frame with 1709 rows and 11 variables:
\describe{
  \item{Feature.ID}{Unique identifier for each record}
  \item{Feature.Name}{Name of building}
  \item{Feature.Class}{Type of building. All records in this dataset will be "hospital".}
  \item{Primary.State}{State of the location}
  \item{Primary.County.Name}{County of the location}
  \item{Elevation..Meters.}{Elevation in meters above sea level of the location}
  \item{Elevation..Feet.}{Elevation in feet above sea level of the location}
  \item{USGS.Map.Name}{Name of the location of the building}
  \item{Date.Created}{Date that the feature was added to the database}
  \item{Date.Edited}{Date at which the record was last updated}
  \item{geometry}{sf point object of geographic location}
}}
\source{
\url{https://data.iowa.gov/Physical-Geography/Iowa-Hospital-Buildings/ciqq-2z9p}
}
\usage{
hospital_buildings
}
\description{
Iowa hospital buildings listed in the federal Geographic Names Information System.
}
\examples{
library(dplyr)

# Leaflet map of hospital buildings in Iowa
library(leaflet)
library(sf)

hospital_buildings \%>\%
 leaflet() \%>\%
 addTiles() \%>\%
 addPolygons(data = st_transform(ia_counties, crs='+proj=longlat +datum=WGS84'),
             weight = 1, color="#333333") \%>\%
 addCircleMarkers(radius = 1, stroke = 0.1)
}
\keyword{datasets}
