% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hospitals}
\alias{hospitals}
\title{Location of hospitals in Iowa}
\format{
A data frame with 145 rows and 15 variables:
\describe{
   \item{city}{name of the city}
   \item{name}{name of the hospital - XXX rename to NAME}
   \item{type}{type of hospital, one of 'critical access', 'general acute care', 'psychiatric', 'long term care', and 'military'}
   \item{website}{url if available}
   \item{beds}{number of BEDS}
   \item{naics_desc}{North American Industry Classification System (NAICS) descriptor: 'general medical and surgical hospital' or 'specialty hospital'}
   \item{result_num}{145 0s XXX delete column?}
   \item{status}{145 OKs XXX delete column?}
   \item{address}{formatted addresses - some are duplicates - XXX look into}
   \item{state}{state abbreviation}
   \item{zip}{zip code}
   \item{latitude}{geographic latitude}
   \item{longitude}{geographic longitude}
   \item{classification}{column used to classify the data set}
}
}
\source{
\url{https://www.officialusa.com/stateguides/health/hospitals/iowa.html}
}
\usage{
hospitals
}
\description{
Dataset was scraped by Masoud Nosrati from Official USA in Mar 2020, geocoding by Andrew Maloney through QGIS.
}
\examples{
# Map of hospitals in Iowa  using ggplot2
library(ggplot2)
library(dplyr) # for the pipe

hospitals \%>\%
  ggplot() +
    geom_point(aes(x = longitude, y = latitude))

# leaflet map
library(leaflet)
library(sf)

hospitals \%>\%
  leaflet() \%>\%
    addTiles() \%>\%
    addPolygons(data = ia_counties,
                weight = 1, color="#333333") \%>\%
    addCircleMarkers(lng = ~longitude, lat = ~latitude,
                     radius = 1, stroke = 0.1,
                     label = ~name,
                     clusterOptions = markerClusterOptions())
}
\keyword{datasets}
