% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ia_counties}
\alias{ia_counties}
\title{Shapefiles of Iowa's counties}
\format{A data frame with 99 rows and 9 variables:
\describe{
  \item{CO_NUMBER}{county number}
  \item{CO_FIPS}{three-digit county fips code}
  \item{ACRES_SF}{square footage in acres}
  \item{ACRES}{county acreage, same as `ACRES_SF``}
  \item{FIPS}{five-digit fips code}
  \item{COUNTY}{county name (and it's `Obrien`)}
  \item{ST}{two letter state abbreviation (`IA` all the way through)}
  \item{ID}{identifier same as `CO_FIPS`}
  \item{CENSUS2010POP}{US Census Bureau count of 2010 county population.}
  \item{POPESTIMATE2019}{US Census Bureau estimate of 2019 county population.}
  \item{geometry}{simple feature object of polygons}
}}
\source{
\url{some url?}
}
\usage{
ia_counties
}
\description{
provenance - Chris, I need some help with how we document these exports.
}
\examples{
# county map of iowa in ggplot2
library(ggplot2)
library(dplyr) # for the pipe

ia_counties \%>\%
  ggplot() +
  geom_sf(aes(fill = POPESTIMATE2019),
          colour = "grey80", size = 0.1) +
  ggthemes::theme_map() +
  theme(legend.position="right") +
  scale_fill_gradient("2019 Population\\nEstimate", trans="log",
    low = "grey80", high="darkred")

# leaflet map
library(leaflet)
library(sf)
st_transform(ia_counties, crs='+proj=longlat +datum=WGS84') \%>\%
  group_by(COUNTY) \%>\%
  mutate(
    hovertext = htmltools::HTML(paste0(COUNTY, "<br>",POPESTIMATE2019))
  ) \%>\%
  leaflet() \%>\%
    addTiles() \%>\%
    addPolygons(label=~hovertext)
}
\keyword{datasets}
