% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ia_education}
\alias{ia_education}
\title{Educational Attainment in Iowa by county}
\format{

}
\source{
https://www.icip.iastate.edu/tables/education/attainment
}
\usage{
ia_education
}
\description{
Percentage of the population ages 25 years or older by highest degree attained.
Data published by ISU extension, based on the 2013 American Community Survey.
}
\keyword{datasets}
