% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ia_election_2016}
\alias{ia_election_2016}
\title{US presidential election 2016 precinct level results for Iowa}
\format{
A data frame with 36,960 rows and 15 variables:
\describe{
  \item{year}{Year of the election.}
  \item{state}{State of election (here, `Iowa`).}
  \item{county_name}{Name of the county.}
  \item{county_fips}{FIPS code of county.}
  \item{county_ansi}{ANSI code of the county.}
  \item{jurisdiction}{County of jurisdiction.}
  \item{precinct}{Precinct name}
  \item{candidate}{Name of the candidate}
  \item{candidate_normalized}{Normalized name of the candidate.}
  \item{writein}{Is candidate a write-in?}
  \item{party}{Party affiliation of the candidate.}
  \item{mode}{How was vote cast? Absentee or at Polling station.}
  \item{votes}{Number of votes cast in precinct.}
  \item{candidate_fec}{Federal Election identifier of the candidate's campaign.}
  \item{candidate_fec_name}{Name associated with the campaign.}
}
}
\source{
\url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/LYWX3D}
}
\usage{
ia_election_2016
}
\description{
Precinct level results from the 2016 general election for all Iowa precincts.
This data was modified from the data provided by MIT Election Data and Science Lab by
focussing on Iowa precinct results from the general presidential election in 2016.
}
\examples{
library(ggplot2)
library(dplyr)

ia_election_2016 \%>\%
  ggplot() +
    geom_bar(aes(x = candidate, weight = votes)) +
    coord_flip()
}
\author{
MIT Election Data and Science Lab (Massachusetts Institute of Technology)
}
\keyword{datasets}
