% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ia_places.R
\docType{data}
\name{ia_places}
\alias{ia_places}
\title{Shapefiles of places in IA in 2020}
\format{
A tibble with 1030 rows and 7 columns
\describe{
  \item{geoid}{seven-digit identifier of the geographic entity}
  \item{name}{character, name of the city. Census designated places show their name with an extension of CDP.  }
  \item{county}{character, name of the (principal) surrounding county}
  \item{county_geoid}{five-digit identifier of the county (FIPS code)}
  \item{county_other}{character, name(s) of the other counties the town/cdp is in.}
  \item{type}{character, classification as "town" or "cdp" (census designated place).}
  \item{ISTP}{logical, is this town part of the Iowa Small Town poll?}
  \item{pop2020}{numeric, 2020 Census Variable P1_001N}
  \item{pop2010}{numeric, 2010 Census Variable P001001}
  \item{pop2000}{numeric, 2000 Census Variable PL001001}
  \item{geometry}{sfc_MULTIPOLYGON, based on the PL 94-171 Redistricting Data Summary File}
}
}
\source{
US Census Bureau Decennial Census 2020, 2010, 2000
}
\usage{
ia_places
}
\description{
The shapefiles are based on places in Iowa in 2020. The names of places and its population numbers are based on
the decennial census counts for 2020, 2010, and 2000.
Note that the US Census Bureau used differential privacy estimates for the 2020 Census, which might impact the
exact numbers below county levels. In particular, for small areas, these estimates might differ from the actual
population numbers.
}
\examples{
# three geometries on top of each other
library(leaflet)
library(leaflet.extras)
library(sf)
  leaflet() \%>\%
  addTiles() \%>\%
  addPolygons(data = ia_counties,
             weight = 1, color="#333333") \%>\%
  addPolygons(data = ia_places, weight = 1, color="maroon", label=~name, group="places") \%>\%
  addSearchFeatures(targetGroups = "places",
                    options = searchFeaturesOptions(
                      zoom = 10, openPopup = TRUE,
                      hideMarkerOnCollapse = TRUE))
}
\keyword{datasets}
