% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ia_tracts}
\alias{ia_tracts}
\title{Shapefiles of Census tracts based on the 2020 Decennial US Census}
\format{
A tibble with 896 rows and 76 columns
\describe{
  \item{GEOID}{character}
  \item{TRACT_ID}{character}
  \item{County_FIPS}{character}
  \item{County}{character}
  \item{P1_001N}{numeric, RACE,  !!Total:}
  \item{P1_002N}{numeric, RACE,  !!Total:!!Population of one race:}
  \item{P1_003N}{numeric, RACE,  !!Total:!!Population of one race:!!White alone}
  \item{P1_004N}{numeric, RACE,  !!Total:!!Population of one race:!!Black or African American alone}
  \item{P1_005N}{numeric, RACE,  !!Total:!!Population of one race:!!American Indian and Alaska Native alone}
  \item{P1_006N}{numeric, RACE,  !!Total:!!Population of one race:!!Asian alone}
  \item{P1_007N}{numeric, RACE,  !!Total:!!Population of one race:!!Native Hawaiian and Other Pacific Islander alone}
  \item{P1_008N}{numeric, RACE,  !!Total:!!Population of one race:!!Some Other Race alone}
  \item{P1_009N}{numeric, RACE,  !!Total:!!Population of two or more races:}
  \item{P1_010N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of two races:}
  \item{P1_011N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of two races:!!White; Black or African American}
  \item{P1_012N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of two races:!!White; American Indian and Alaska Native}
  \item{P1_013N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of two races:!!White; Asian}
  \item{P1_014N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of two races:!!White; Native Hawaiian and Other Pacific Islander}
  \item{P1_015N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of two races:!!White; Some Other Race}
  \item{P1_016N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of two races:!!Black or African American; American Indian and Alaska Native}
  \item{P1_017N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of two races:!!Black or African American; Asian}
  \item{P1_018N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of two races:!!Black or African American; Native Hawaiian and Other Pacific Islander}
  \item{P1_019N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of two races:!!Black or African American; Some Other Race}
  \item{P1_020N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of two races:!!American Indian and Alaska Native; Asian}
  \item{P1_021N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of two races:!!American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
  \item{P1_022N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of two races:!!American Indian and Alaska Native; Some Other Race}
  \item{P1_023N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of two races:!!Asian; Native Hawaiian and Other Pacific Islander}
  \item{P1_024N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of two races:!!Asian; Some Other Race}
  \item{P1_025N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of two races:!!Native Hawaiian and Other Pacific Islander; Some Other Race}
  \item{P1_026N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:}
  \item{P1_027N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:!!White; Black or African American; American Indian and Alaska Native}
  \item{P1_028N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:!!White; Black or African American; Asian}
  \item{P1_029N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:!!White; Black or African American; Native Hawaiian and Other Pacific Islander}
  \item{P1_030N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:!!White; Black or African American; Some Other Race}
  \item{P1_031N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:!!White; American Indian and Alaska Native; Asian}
  \item{P1_032N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:!!White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
  \item{P1_033N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:!!White; American Indian and Alaska Native; Some Other Race}
  \item{P1_034N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:!!White; Asian; Native Hawaiian and Other Pacific Islander}
  \item{P1_035N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:!!White; Asian; Some Other Race}
  \item{P1_036N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:!!White; Native Hawaiian and Other Pacific Islander; Some Other Race}
  \item{P1_037N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:!!Black or African American; American Indian and Alaska Native; Asian}
  \item{P1_038N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:!!Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
  \item{P1_039N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:!!Black or African American; American Indian and Alaska Native; Some Other Race}
  \item{P1_040N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:!!Black or African American; Asian; Native Hawaiian and Other Pacific Islander}
  \item{P1_041N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:!!Black or African American; Asian; Some Other Race}
  \item{P1_042N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:!!Black or African American; Native Hawaiian and Other Pacific Islander; Some Other Race}
  \item{P1_043N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:!!American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
  \item{P1_044N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:!!American Indian and Alaska Native; Asian; Some Other Race}
  \item{P1_045N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:!!American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race}
  \item{P1_046N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of three races:!!Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
  \item{P1_047N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of four races:}
  \item{P1_048N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of four races:!!White; Black or African American; American Indian and Alaska Native; Asian}
  \item{P1_049N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of four races:!!White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander}
  \item{P1_050N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of four races:!!White; Black or African American; American Indian and Alaska Native; Some Other Race}
  \item{P1_051N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of four races:!!White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander}
  \item{P1_052N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of four races:!!White; Black or African American; Asian; Some Other Race}
  \item{P1_053N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of four races:!!White; Black or African American; Native Hawaiian and Other Pacific Islander; Some Other Race}
  \item{P1_054N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of four races:!!White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
  \item{P1_055N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of four races:!!White; American Indian and Alaska Native; Asian; Some Other Race}
  \item{P1_056N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of four races:!!White; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race}
  \item{P1_057N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of four races:!!White; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
  \item{P1_058N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of four races:!!Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
  \item{P1_059N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of four races:!!Black or African American; American Indian and Alaska Native; Asian; Some Other Race}
  \item{P1_060N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of four races:!!Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race}
  \item{P1_061N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of four races:!!Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
  \item{P1_062N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of four races:!!American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
  \item{P1_063N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of five races:}
  \item{P1_064N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of five races:!!White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander}
  \item{P1_065N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of five races:!!White; Black or African American; American Indian and Alaska Native; Asian; Some Other Race}
  \item{P1_066N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of five races:!!White; Black or African American; American Indian and Alaska Native; Native Hawaiian and Other Pacific Islander; Some Other Race}
  \item{P1_067N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of five races:!!White; Black or African American; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
  \item{P1_068N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of five races:!!White; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
  \item{P1_069N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of five races:!!Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
  \item{P1_070N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of six races:}
  \item{P1_071N}{numeric, RACE,  !!Total:!!Population of two or more races:!!Population of six races:!!White; Black or African American; American Indian and Alaska Native; Asian; Native Hawaiian and Other Pacific Islander; Some Other Race}
  \item{geometry}{sfc_POLYGON}
}
}
\usage{
ia_tracts
}
\description{
Shapefiles for Census tracts areas are made available by the US Census Bureau through the Census API.
}
\keyword{datasets}
