% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iowa_features}
\alias{iowa_features}
\title{Physical and Cultural Geographic Features in Iowa}
\format{A data frame with 40,473 rows and 11 variables:
\describe{
  \item{Feature.ID}{Permanent, unique feature record identifier as defined in INCITS 446-2008. }
  \item{Feature.Name}{Official feature name as defined in INCITS 446-2008.}
  \item{Feature.Class}{Type of feature, one of Airport, Arch, Area, Bar, Bay, Beach, Bend, Bridge, Building, Canal, Cape, Cemetery, Census, Channel, Church, Civil, Cliff, Crossing, Dam, Falls, Flat, Forest, Gut, Harbor, Hospital, Island, Lake, Levee, Locale, Military, Mine, Oilfield, Park, Pillar, Plain, Populated Place, Post Office, Range, Reserve, Reservoir, Ridge, School, Spring, Stream, Summit, Swamp, Tower, Trail, Tunnel, Valley, and Woods..}
  \item{Primary.County.Name}{County of the primary location of the feature.}
  \item{Primary.Point..Coordinates.}{just kept for now, will be deleted XXX}
  \item{Primary.State}{Two-letter state abbreviation. Constant value of `IA`.}
  \item{Elevation.M}{Elevation in meters above sea level of the feature.}
  \item{Elevation.Ft}{Elevation in feet above sea level of the feature.}
  \item{USGS.Map.Name}{Name of the location of the building}
  \item{Date.Created}{Date that the feature was added to the database}
  \item{Date.Edited}{Date at which the record was last updated}
  \item{geometry}{sf point object of geographic location}
}}
\source{
\url{https://data.iowa.gov/Physical-Geography/Iowa-Physical-and-Cultural-Geographic-Features/uedc-2fk7}
}
\usage{
iowa_features
}
\description{
This dataset consists of all named physical and cultural geographic features in Iowa that are part of the
Geographic Names Information System (GNIS) as developed by the US Geological Survey.
The data was downloaded from the Iowa Data Portal in July 2020.
It encompasses various aspects of Iowa's infrastructure, including hospital buildings and parks.
}
\examples{
library(dplyr)
library(ggplot2)
iowa_features \%>\%
  filter(Feature.Class == "Park") \%>\%
  ggplot() +
  geom_sf() +
  xlim(c(-97,-89.75)) +
  ylim(c(40, 45))

# The Bridges of Madison County
  library(leaflet)
  library(sf)

 iowa_features \%>\%
   filter(Primary.County.Name == "Madison",
          Feature.Class == "Bridge") \%>\%
   leaflet() \%>\%
   addTiles() \%>\%
   setView(-94.0530854, 41.3409121, zoom = 9) \%>\%
   addPolygons(data = ia_counties,
               weight = 1, color="#333333") \%>\%
   addCircleMarkers(radius = 1, stroke = 0.1, label=~Feature.Name)

# Leaflet map of top ten feature maps in Iowa
 topten <- iowa_features \%>\% count(Feature.Class) \%>\%
             arrange(desc(n)) \%>\% head(10)

 pal <- colorFactor(
          palette=RColorBrewer::brewer.pal(n=10, name="Paired"),
          domain = topten$Feature.Class
        )
 iowa_features \%>\%
   filter(Feature.Class \%in\% topten$Feature.Class) \%>\%
   leaflet() \%>\%
   addTiles() \%>\%
   setView(-94, 42, zoom = 6) \%>\%
   addPolygons(data = ia_counties,
               weight = 1, color="#333333") \%>\%
   addCircleMarkers(radius = .5, stroke = 0.1,
                    color =~pal(Feature.Class),
                    label=~Feature.Name) \%>\%
   addLegend(pal = pal, values = topten$Feature.Class)

 # Iowa is flat. Is it? Yes, but ...
 elev_pal <- colorNumeric(
          palette="Blues",
          domain = c(0,600),
          reverse=TRUE
        )

 iowa_features \%>\%
   filter(Feature.Class \%in\% topten$Feature.Class) \%>\%
   leaflet() \%>\%
   addTiles() \%>\%
   setView(-94, 42, zoom = 6) \%>\%
   addPolygons(data = ia_counties,
               weight = 1, color="#333333") \%>\%
   addCircleMarkers(radius = .5, stroke = 0.1,
                    color =~elev_pal(Elevation.M),
                    label=~Feature.Name) \%>\%
   addLegend(pal=elev_pal, values=c(0,600))
}
\keyword{datasets}
