% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iowaworks}
\alias{iowaworks}
\title{Location of IowaWORKS Centers in Iowa}
\format{A data frame with 27 rows and 18 variables:
\describe{
   \item{ID}{identifier, not quite the row number}
   \item{NAME}{enumeration of counties the center serves}
   \item{TYPE}{Center - XXX delete}
   \item{STREET}{street address}
   \item{CITY}{city}
   \item{STATE}{state}
   \item{ZIP}{5-digit zip code}
   \item{PHONE}{phone number}
   \item{EMAIL}{email address}
   \item{AREA}{NAs - XXX delete}
   \item{LINK}{weblink to the center}
   \item{result_num}{0s XXX delete column?}
   \item{status}{OKs XXX delete column?}
   \item{formatted_}{formatted addresses - some are duplicates - XXX look into}
   \item{place_id}{identifier, based on address? - some are duplicates - XXX look into}
   \item{location_t}{categorical variable with additional details on location.}
   \item{Latitude}{geographic latitude}
   \item{Longitude}{geographic longitude}
}}
\source{
\url{https://www.iowaworkforcedevelopment.gov/contact}
}
\usage{
iowaworks
}
\description{
Dataset was scraped by Masoud Nosrati from the Iowa Workforce Development in Mar 2020, geocoding by Andrew Maloney through QGIS.
}
\examples{
# Map of centers in Iowa  using ggplot2
library(ggplot2)
library(dplyr) # for the pipe

iowaworks \%>\%
  ggplot() +
    geom_point(aes(x = Longitude, y = Latitude))

# leaflet map
library(leaflet)
library(sf)

iowaworks \%>\%
  group_by(NAME, PHONE) \%>\%
  mutate(
    hovertext = htmltools::HTML(paste0("IowaWORKS Center<br>",formatted_, '<br>serving ', NAME, '<br>', PHONE))
  ) \%>\%
  leaflet() \%>\%
    addTiles() \%>\%
    addPolygons(data = st_transform(ia_counties, crs='+proj=longlat +datum=WGS84'),
                weight = 1, color="#333333") \%>\%
    addCircleMarkers(lng = ~Longitude, lat = ~Latitude,
                     radius = 1, stroke = 0.1,
                     label = ~hovertext)
}
\keyword{datasets}
