% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mat}
\alias{mat}
\title{Location of MAT in Iowa}
\format{A data frame with 145 rows and 15 variables:
\describe{
   \item{ID}{identifier, not quite the row number}
   \item{DOCTOR}{name of the medical personnel}
   \item{CENTER}{MAT center}
   \item{STREET}{street address}
   \item{CITY}{city}
   \item{STATE}{state}
   \item{ZIP}{5-digit zip code}
   \item{PHONE}{phone number}
   \item{TREATMENT}{treatment options: methadone and/or buprenorphine}
   \item{result_num}{0s XXX delete column?}
   \item{status}{OKs XXX delete column?}
   \item{formatted_}{formatted addresses - some are duplicates - XXX look into}
   \item{place_id}{identifier, based on address? - some are duplicates - XXX look into}
   \item{location_t}{categorical variable with additional details on location.}
   \item{Latitude}{geographic latitude}
   \item{Longitude}{geographic longitude}
}}
\source{
\url{https://iowa.maps.arcgis.com/apps/LocalPerspective/index.html?appid=924e0f99711b406dbf22a34cf46fc6e1}
}
\usage{
mat
}
\description{
This dataset provides doctors available in Medication Assisted Treatment (MAT) facilities in Iowa.
Dataset was scraped by Masoud Nosrati from the Iowa Department of Public Health in Mar 2020, geocoding by Andrew Maloney through QGIS.
}
\examples{
# Map of hospitals in Iowa  using ggplot2
library(ggplot2)
library(dplyr) # for the pipe

mat \%>\%
  ggplot() +
    geom_point(aes(x = Longitude, y = Latitude))

# leaflet map
library(leaflet)
library(sf)

mat \%>\%
  group_by(DOCTOR, CENTER) \%>\%
  mutate(
    hovertext = htmltools::HTML(paste0(DOCTOR, '<br>', CENTER, '<br>', PHONE))
  ) \%>\%
  leaflet() \%>\%
    addTiles() \%>\%
    addPolygons(data = ia_counties,
                weight = 1, color="#333333") \%>\%
    addCircleMarkers(lng = ~Longitude, lat = ~Latitude,
                     radius = 1, stroke = 0.1,
                     label = ~hovertext,
                     clusterOptions = markerClusterOptions())
}
\keyword{datasets}
