% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meetings}
\alias{meetings}
\title{Alcoholic and Narcotics Anonymous Meetings in Iowa}
\format{
A data frame with 1549 rows and 13 variables:
\describe{
  \item{day}{day of the week}
  \item{time}{time of the day}
  \item{ampm}{morning or afternoon}
  \item{meeting}{name of the meeting}
  \item{address}{address}
  \item{street}{stree name}
  \item{state}{state}
  \item{format}{Format of the meeting - XXX this needs some additional work}
  \item{city}{city}
  \item{type}{type of meeting}
  \item{longitude}{geographic Longitude}
  \item{latitude}{geographic Latitude}
  \item{schedule}{weekly schedule of meetings, starting with Monday at midnight.}
  \item{classification}{column used to classify the data set}
  \item{county}{county}
}
}
\source{
\url{https://www.aa-iowa.org/meetings/}, \url{https://www.na-iowa.org/meetings/},
\url{https://adultchildren.org/mtsearch}, \url{https://al-anon.org/al-anon-meetings/find-an-alateen-meeting/},
\url{http://draonline.qwknetllc.com/meetings_dra/usa/iowa.html}, \url{https://www.nar-anon.org},
\url{https://www.smartrecoverytest.org/local/full-meeting-list-download/}, \url{https://locator.crgroups.info/},
\url{https://www.facebook.com/crushofiowa/}, \url{https://refugerecovery.org/meetings?tsml-day=any&tsml-region=iowa}
}
\usage{
meetings
}
\description{
Dataset was scraped from the support group websites of Iowa by
Jessie Bustin, Grant Durhban, Atefeh Rajabalizadeh, and Vikram Magal.
}
\examples{
# Location of meetings in Iowa
library(ggplot2)
library(dplyr)
meetings \%>\%
  ggplot(aes(x = longitude, y = latitude)) +
  geom_point(aes(colour = type))

# Leaflet map of meetings in Iowa
library(leaflet)
library(sf)
meetings \%>\%
 leaflet() \%>\%
 addTiles() \%>\%
 addPolygons(data = ia_counties,
             weight = 1, color="#333333") \%>\%
 addCircleMarkers(meetings \%>\% filter(type == "Narcotics Anonymous"),
                  lng = ~longitude, lat = ~latitude,
                  radius = 1, stroke = 0.1,
                  color = "darkorange",
                  label = ~meeting, group="NA") \%>\%
 addCircleMarkers(meetings \%>\% filter(type == "Alcoholics Anonymous"),
                  lng = ~longitude, lat = ~latitude,
                  radius = 1, stroke = 0.1,
                  color = "darkgreen",
                  label = ~meeting, group="AA") \%>\%
 addLayersControl(overlayGroups = c("AA","NA"))
}
\keyword{datasets}
