% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recovery}
\alias{recovery}
\title{Recovery data}
\format{
A tibble with 156 rows and 42 columns
\describe{
  \item{IDPH_Feedback}{character}
  \item{Name}{character}
  \item{Street}{character}
  \item{City}{character}
  \item{County}{character}
  \item{State}{character}
  \item{Zip}{character}
  \item{Phone}{character}
  \item{Email}{character}
  \item{URL}{character}
  \item{RecoveryHousing}{character}
  \item{RecoveryServices}{character}
  \item{Type of structure}{character}
  \item{twelve_Step}{character}
  \item{Residential}{character}
  \item{Groups}{character}
  \item{InpatientTreatment}{character}
  \item{CaseManagement}{character}
  \item{DBT}{character}
  \item{CBT}{character}
  \item{REBT}{character}
  \item{Intensive_Outpatient}{character}
  \item{OutpatientTreatment}{character}
  \item{Medications}{character}
  \item{Trauma_PTSD}{character}
  \item{Family}{character}
  \item{Veterans}{character}
  \item{ReligiousBased}{character}
  \item{Gender}{character}
  \item{Services offered}{character}
  \item{Male}{character}
  \item{Female}{character}
  \item{Families}{character}
  \item{Minors}{character}
  \item{Pets}{character}
  \item{Hours}{character}
  \item{Days}{character}
  \item{Fees}{character}
  \item{Sources (Page Links)}{character}
  \item{Address}{character}
  \item{lon}{numeric}
  \item{lat}{numeric}
}
}
\usage{
recovery
}
\description{
Dataset formerly used as Reco-data.csv
}
\keyword{datasets}
