% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regional_substance_treatment}
\alias{regional_substance_treatment}
\title{Comprehensive Substance Abuse Prevention Grant Facilities}
\format{
A data frame with 129 rows and 11 variables:
\describe{
  \item{classification}{The systems of care infrastructure category under which the facility falls; all in this dataset are considered substance use treatment}
  \item{dataset}{For possible future use, all data are in the regional_substance_treatment dataset}
  \item{regional_org}{The regional organization that manages the specific facility in the observation}
  \item{source}{The link from which the data was obtained}
  \item{location_name}{The name of the specific treatment facility}
  \item{address}{The street address of the treatment facility}
  \item{unit_num}{The unit or suite number of the office, if it has one}
  \item{city}{The city of the treatment facility}
  \item{state}{The state of the treatment facility}
  \item{zip}{The zip code of the treatment facility}
  \item{phone}{The phone number for the treatment facility}
}
}
\source{

}
\usage{
regional_substance_treatment
}
\description{
This dataset includes all facilities managed under the Iowa Comprehensive Substance Abuse Prevention Grant regional centers. These 19 regional organizations manage prevention, inpatient, and outpatient substance use facilities in their respective regions of Iowa. The data was collected by Matthew Voss in July 2020.
}
\keyword{datasets}
