% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rx_Drop_Off_Locations}
\alias{Rx_Drop_Off_Locations}
\title{Rx drop off locations}
\format{
A data frame with 312 rows and 12 variables:
\describe{
  \item{name}{Name of the accesss point}
  \item{address}{Address of the building}
  \item{city}{City name}
  \item{state}{State name}
  \item{zip}{Zip code}
  \item{type}{Type of facility}
  \item{phone}{Phone number of the building}
  \item{objectid}{Object ID}
  \item{x}{Utm co-ordinate}
  \item{y}{Utm co-ordinate}
  \item{dataset}{Dataset name}
  \item{classification}{domain}
  \item{search_address}{Full address}
  \item{longitude}{geographic Latitude}
  \item{latitude}{geographic Longitude}
}
}
\source{
\url{https://iowa.maps.arcgis.com/apps/LocalPerspective/index.html?appid=924e0f99711b406dbf22a34cf46fc6e1}
}
\usage{
Rx_Drop_Off_Locations
}
\description{
Rx Drop Off Locations across the state by type and county
}
\keyword{datasets}
