% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{southwest_mhds}
\alias{southwest_mhds}
\title{Southwest Iowa Region Mental Health & Disability Services}
\format{A data frame with 356 rows and 13 variables:
\describe{
  \item{Category}{Service type provided at facility}
  \item{County}{Iowa county name}
  \item{Related Records}{Currently Unknown}
  \item{Service Title}{Name of the facility}
  \item{Address}{Column containing location information: Street, City, State, Zip}
  \item{Phone}{Phone Number for the facility}
  \item{Description}{A description to help the viewer navigate which facility is best for them}
  \item{Website}{Facility website link}
  \item{Email}{Contact Email for facility}
  \item{Expiration}{Currently Unknown}
  \item{Last Updated}{Date at which website information was last updated}
  \item{lon}{geographic Longitude}
  \item{lat}{geographic Latitude}
}}
\source{
\url{https://www.aa-iowa.org/meetings/}
}
\usage{
southwest_mhds
}
\description{
Dataset was scraped from the MHDS website.
}
\examples{
# Location of MHDS in Iowa
library(ggplot2)
library(dplyr)
southwest_mhds \%>\%
 ggplot() +
 geom_point(aes(x = lon, y = lat))

# Leaflet map of meetings in Iowa
library(leaflet)
library(sf)

southwest_mhds \%>\%
 leaflet() \%>\%
 addTiles() \%>\%
 addPolygons(data = ia_counties,
             weight = 1, color="#333333") \%>\%
 addCircleMarkers(lng = ~lon, lat = ~lat,
                  radius = 1, stroke = 0.1,
                  label = ~southwest_mhds)
}
\keyword{datasets}
