% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{specialty_courts}
\alias{specialty_courts}
\title{Specialty Courts in Iowa}
\format{
A data frame with 47 rows and 5 variables:
\describe{
   \item{dataset}{name of the dataset}
   \item{classification}{type of services offered by datset}
   \item{id}{identification number, not quite row number}
   \item{court_type}{denotes what specialty court is for}
   \item{county}{county of the court}
}
}
\source{
\url{https://www.iowacourts.gov/media/cms/Iowa_Specialty_Courts_8C3B9AF1F3CA7.pdf}
}
\usage{
specialty_courts
}
\description{
This dataset was scraped by Masoud Nosrati in July 2020 from wikipedia.
The data includes hospitals in Iowa scraped from Wikipedia
}
\keyword{datasets}
