% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sud}
\alias{sud}
\title{Location of Licensed Substance Use Disorder/Problem Gambling Programs in Iowa}
\format{
A data frame with 123 rows and 15 variables:
\describe{
   \item{name}{name of the facility}
   \item{phone}{phone number}
   \item{street}{street address}
   \item{city}{city}
   \item{statename}{name of state}
   \item{zip}{5-digit zip code}
   \item{result_num}{0s XXX delete column?}
   \item{status}{OKs XXX delete column?}
   \item{address}{formatted addresses - some are duplicates - XXX look into}
   \item{latitude}{geographic latitude}
   \item{longitude}{geographic longitude}
   \item{state}{state abbreviation}
   \item{classification}{column used to classify the dataset}
}
}
\source{
\url{https://www.iowaworkforcedevelopment.gov/contact}
}
\usage{
sud
}
\description{
Dataset was scraped by Masoud Nosrati from the Iowa Department of Public Health Development in Mar 2020, geocoding by Andrew Maloney through QGIS.
}
\examples{
# Map of centers in Iowa  using ggplot2
library(ggplot2)
library(dplyr) # for the pipe

sud \%>\%
  ggplot() +
    geom_point(aes(x = longitude, y = latitude))

# leaflet map
library(leaflet)
library(sf)

sud \%>\%
  group_by(name,address, phone) \%>\%
  mutate(
    hovertext = htmltools::HTML(paste0(name, "<br>",formatted_, '<br>', phone))
  ) \%>\%
  leaflet() \%>\%
    addTiles() \%>\%
    addPolygons(data = ia_counties,
                weight = 1, color="#333333") \%>\%
    addCircleMarkers(lng = ~longitude, lat = ~latitude,
                     radius = 1, stroke = 0.1,
                     label = ~hovertext)
}
\keyword{datasets}
