% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_hospitals}
\alias{wiki_hospitals}
\title{Hospitals in Iowa}
\format{
A data frame with 123 rows and 6 variables:
\describe{
   \item{dataset}{name of the dataset}
   \item{classification}{type of services offered by datset}
   \item{name}{name of hospital}
   \item{city}{city hospital is located in}
   \item{county}{county of the hospital}
   \item{state}{state of the hospital}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_hospitals_in_Iowa}
}
\usage{
wiki_hospitals
}
\description{
This dataset was scraped by Masoud Nosrati in July 2020 from wikipedia.
The data includes hospitals in Iowa scraped from Wikipedia
}
\keyword{datasets}
