% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{workforce_development}
\alias{workforce_development}
\title{Workforce development centers in Iowa}
\format{
A data frame with 27 rows and 16 variables:
\describe{
   \item{dataset}{name of the dataset}
   \item{classification}{type of services offered by datset}
   \item{id}{identification number, not quite row number}
   \item{counties_served}{Counties serviced by the center}
   \item{type}{type of building}
   \item{street}{street address of the center}
   \item{city}{city in which the center is located}
   \item{state}{state in which the center is located}
   \item{zip}{zip code of the center}
   \item{county}{county in which the center is located}
   \item{phone}{phone number of the center}
   \item{email}{email of the center}
   \item{link}{link to center's website}
   \item{address}{search address used in geocoding}
   \item{longitude}{longitude coordinate}
   \item{latitude}{latitude coordinate}
}
}
\source{
\url{https://www.iowaworkforcedevelopment.gov/contact}
}
\usage{
workforce_development
}
\description{
This dataset contains data on workforce development centers in Iowa.
Data was scraped by Masoud Nosrati in July 2020 and geocoded by Joel Von Behren in July 2020.
}
\keyword{datasets}
