% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zip10_ia}
\alias{zip10_ia}
\title{Zipcode shapefiles based on the 2010 US Census}
\format{
A tibble with 934 rows and 10 columns
\describe{
  \item{ZCTA5CE10}{character, 2010 Census 5-digit ZIP Code Tabulation Area code}
  \item{GEOID10}{character, 2010 Census 5-digit ZIP Code Tabulation Area identifier, 2010 Census 5-digit ZIP Code Tabulation Area code}
  \item{city}{character, associated city.}
  \item{county}{character, county of the zipcode.}
  \item{type}{character, one of "Standard", "P.O. Box", or "Unique".}
  \item{ALAND10}{numeric, 2010 Census land area (square meters)}
  \item{AWATER10}{numeric, 2010 Census water area (square meters)}
  \item{INTPTLAT10}{numeric, 2010 Census latitude of the internal point}
  \item{INTPTLON10}{numeric, 2010 Census longitude of the internal point}
  \item{geometry}{sfc_MULTIPOLYGON}
}
}
\usage{
zip10_ia
}
\description{
Shapefiles for zipcode areas are made available by the US Census Bureau
through the data catalog at \url{https://catalog.data.gov/dataset/tiger-line-shapefile-2019-2010-nation-u-s-2010-census-5-digit-zip-code-tabulation-area-zcta5-na}.
Additional information on zipcode areas, such as City, County, and Type of zipcode is added
from zip-codes.com.
}
\keyword{datasets}
