% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs}
\alias{acs}
\title{American Community Survey - Computer Presence}
\format{A data frame with 99 rows and 9 variables:
\describe{
  \item{Geography.Id}{geographic ID used by the U.S. census for the location associated with this record}
  \item{Type}{The type of location associated with this record. Categories are state, county, place, and tract}
  \item{Name}{Name of the location}
  \item{Variable}{Variable ID identified by the U.S. Census Bureau}
  \item{Variable.Description}{Description of the variable}
  \item{Computer.Present}{Classifies if a computer is present in those households. Categories are total, yes, and no}
  \item{Internet.Subscription}{Classifies what kind of internet subscription is present in those households. Categories are total, total w/computer, broadband, dial-up, and none}
  \item{Data.Collection.Period}{Period in which the data was collected}
  \item{Data.Collection.End.Date}{The date in which the data was done being collected}
  \item{Households}{The number of households estimated to have the specified characteristics in the record}
  \item{Row.ID}{Unique ID associated with the record}
  \item{geometry}{sf point object of geographic location}
}}
\source{
\url{https://data.iowa.gov/Utilities-Telecommunications/Iowa-Households-by-Presence-of-a-Computer-and-Type/gz3j-hzab}
}
\usage{
acs
}
\description{
Summary data of Iowa American Community Survey responses by computer presence and internet subscription status. Each row represents a combination of variables and the households variable representing the number of households estimated to be associated with that variable combination.
}
\examples{
# county map of iowa in ggplot2
library(ggplot2)
library(dplyr) # for the pipe

ia_counties \%>\%
  ggplot() + geom_sf(aes(fill = ACRES_SF)) +
  ggthemes::theme_map()

# leaflet map
library(leaflet)
library(sf)
st_transform(ia_counties, crs='+proj=longlat +datum=WGS84') \%>\%
  leaflet() \%>\%
    addTiles() \%>\%
    addPolygons()
}
\keyword{datasets}
