% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asac_locations}
\alias{asac_locations}
\title{Location of ASAC (Area Substance Abuse Counseling) in Iowa}
\format{A data frame with 15 rows and 9 variables:
\describe{
  \item{Name}{Name of the facility}
  \item{Address}{Street Address for the facility}
  \item{City_State_Zip}{City, State, and Zip contained in one column separated by a space}
  \item{City}{Name of the city where the facility is located}
  \item{Zip}{5 digit Zip}
  \item{State}{State Abbreviation}
  \item{search_address}{address used in the Google API geocoding}
  \item{lon}{geographic Longitude}
  \item{lat}{geographic Latitude}
}}
\source{
\url{http://www.asac.us/about/locations/}
}
\usage{
asac_locations
}
\description{
Dataset was scraped from the ASAC website.
This database contains ASAC location data for the State of Iowa
}
\examples{
library(ggplot2)
library(dplyr)
asac_locations \%>\%
 ggplot() +
 geom_point(aes(x = lon, y = lat))

library(leaflet)
library(sf)
asac_locations \%>\%
 leaflet() \%>\%
 addTiles() \%>\%
 addPolygons(data = st_transform(ia_counties, crs='+proj=longlat +datum=WGS84'),
             weight = 1, color="#333333") \%>\%
 addCircleMarkers(lng = ~lon, lat = ~lat,
                  radius = 1, stroke = 0.1)
}
\keyword{datasets}
