% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asac_locations}
\alias{asac_locations}
\title{Location of ASAC (Area Substance Abuse Counseling) in Iowa}
\format{An object of class \code{data.frame} with 15 rows and 9 columns.}
\source{
\url{http://www.asac.us/about/locations/}
}
\usage{
asac_locations
}
\description{
Dataset was scraped from the ASAC website.
This database contains ASAC location data for the State of Iowa
\describe{
  \item{name}{Name of the facility}
  \item{address}{Street Address for the facility}
  \item{city_state_zip}{City, State, and Zip contained in one column separated by a space}
  \item{city}{Name of the city where the facility is located}
  \item{zip}{5 digit Zip}
  \item{state}{State Abbreviation}
  \item{search_address}{address used in the Google API geocoding}
  \item{geometry}{sf object of geographic locations}
  \item{classification}{column used for identification for filtering}
}
}
\examples{
library(ggplot2)
library(dplyr)
asac_locations \%>\%
 ggplot() +
 geom_sf(data = ia_counties) + # iowa county shapes
 geom_sf() # add points

library(leaflet)
library(sf)
asac_locations \%>\%
 leaflet() \%>\%
 addTiles() \%>\%
 addPolygons(data = ia_counties,
             weight = 1, color="#333333") \%>\%
 addCircleMarkers(radius = 1, stroke = 0.1, label=~Name)
}
\keyword{datasets}
