% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{churches}
\alias{churches}
\title{Location of churches in Iowa}
\format{A data frame with 5469 rows and 9 variables:
\describe{
   \item{name}{name of the church}
   \item{description}{closer description of the location, if available}
   \item{county}{name of the county}
   \item{geometry}{sf point object of geographic locations of churches in Iowa.}
   \item{classification}{column used for classifying the data set}
}}
\source{
\url{https://iowa.hometownlocator.com/features/cultural,class,church.cfm}
}
\usage{
churches
}
\description{
Dataset was scraped by Masoud Nosrati from the IA Hometown Locator in Mar 2020.
}
\examples{
# Map of churches in Story county using ggplot2
library(ggplot2)
library(dplyr) # for the pipe

churches \%>\%
 # filter(COUNTY == "Story county) \%>\%
  ggplot() +
    geom_sf(data = ia_counties) +
    geom_sf() +
    ggthemes::theme_map()

# leaflet map
library(leaflet)
library(sf)

churches \%>\%
  leaflet() \%>\%
    addTiles() \%>\%
    setView(-93.6498803, 42.0275751, zoom = 8) \%>\%
    addPolygons(data = ia_counties,
                weight = 1, color="#333333") \%>\%
    addCircleMarkers(radius = 1, stroke = 0.1,
                     label = ~NAME, clusterOptions = markerClusterOptions())
}
\keyword{datasets}
