% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hospital_buildings}
\alias{hospital_buildings}
\title{Iowa Hospital Buildings}
\format{A data frame with 1709 rows and 11 variables:
\describe{
  \item{name}{Name of building}
  \item{feature_class}{Type of building. All records in this dataset will be "hospital".}
  \item{state}{State of the location}
  \item{county}{County of the location}
  \item{elevation_meters}{Elevation in meters above sea level of the location}
  \item{elevation__feet}{Elevation in feet above sea level of the location}
  \item{usgs_map_name}{Name of the location of the building}
  \item{date_created}{Date that the feature was added to the database}
  \item{date_edited}{Date at which the record was last updated}
  \item{geometry}{sf point object of geographic location}
  \item{classification}{column used for classifying the data set}
}}
\source{
\url{https://data.iowa.gov/Physical-Geography/Iowa-Hospital-Buildings/ciqq-2z9p}
}
\usage{
hospital_buildings
}
\description{
Iowa hospital buildings listed in the federal Geographic Names Information System.
}
\examples{
library(dplyr)

# Leaflet map of hospital buildings in Iowa
library(leaflet)
library(sf)

hospital_buildings \%>\%
 leaflet() \%>\%
 addTiles() \%>\%
 addPolygons(data = ia_counties,
             weight = 1, color="#333333") \%>\%
 addCircleMarkers(radius = 1, stroke = 0.1)
}
\keyword{datasets}
