% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hospitals}
\alias{hospitals}
\title{Location of hospitals in Iowa}
\format{A data frame with 145 rows and 15 variables:
\describe{
   \item{ID}{identifier, not quite the row number}
   \item{City}{name of the city}
   \item{Hospital.N}{name of the hospital - XXX rename to NAME}
   \item{TYPE}{type of hospital, one of 'critical access', 'general acute care', 'psychiatric', 'long term care', and 'military'}
   \item{WEBSITE}{url if available}
   \item{ADDRESS}{address of the health clinic}
   \item{BEDS}{number of BEDS}
   \item{NAICS_DESC}{North American Industry Classification System (NAICS) descriptor: 'general medical and surgical hospital' or 'specialty hospital'}
   \item{result_num}{145 0s XXX delete column?}
   \item{status}{145 OKs XXX delete column?}
   \item{formatted_}{formatted addresses - some are duplicates - XXX look into}
   \item{place_id}{identifier, based on address? - some are duplicates - XXX look into}
   \item{location_t}{categorical variable with additional details on location.}
   \item{Latitude}{geographic latitude}
   \item{Longitude}{geographic longitude}
}}
\source{
\url{https://www.officialusa.com/stateguides/health/hospitals/iowa.html}
}
\usage{
hospitals
}
\description{
Dataset was scraped by Masoud Nosrati from Official USA in Mar 2020, geocoding by Andrew Maloney through QGIS.
}
\examples{
# Map of hospitals in Iowa  using ggplot2
library(ggplot2)
library(dplyr) # for the pipe

hospitals \%>\%
  ggplot() +
    geom_point(aes(x = Longitude, y = Latitude))

# leaflet map
library(leaflet)
library(sf)

hospitals \%>\%
  leaflet() \%>\%
    addTiles() \%>\%
    addPolygons(data = ia_counties,
                weight = 1, color="#333333") \%>\%
    addCircleMarkers(lng = ~Longitude, lat = ~Latitude,
                     radius = 1, stroke = 0.1,
                     label = ~Hospital.N,
                     clusterOptions = markerClusterOptions())
}
\keyword{datasets}
