% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meetings}
\alias{meetings}
\title{AA Meetings in Iowa}
\format{A data frame with 15 rows and 9 variables:
\describe{
  \item{Day}{day of the week}
  \item{Time}{time of the day}
  \item{AmPm}{morning or afternoon}
  \item{Meeting}{name of the meeting}
  \item{Location}{location}
  \item{Address}{address}
  \item{Types}{Types of the meeting - XXX this needs some additional work}
  \item{City}{city}
  \item{Type}{type of meeting}
  \item{State}{state}
  \item{lon}{geographic Longitude}
  \item{lat}{geographic Latitude}
}}
\source{
\url{https://www.aa-iowa.org/meetings/}
}
\usage{
meetings
}
\description{
Dataset was scraped from the AA website by Jessie Bustin.
}
\examples{
# Location of meetings in Iowa
library(ggplot2)
library(dplyr)
meetings \%>\%
 ggplot() +
 geom_point(aes(x = lon, y = lat))

# Leaflet map of meetings in Iowa
library(leaflet)
library(sf)
meetings \%>\%
 leaflet() \%>\%
 addTiles() \%>\%
 addPolygons(data = st_transform(ia_counties, crs='+proj=longlat +datum=WGS84'),
             weight = 1, color="#333333") \%>\%
 addCircleMarkers(lng = ~lon, lat = ~lat,
                  radius = 1, stroke = 0.1,
                  label = ~Meeting)
}
\keyword{datasets}
