% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meetings}
\alias{meetings}
\title{Alcoholic and Narcotics Anonymous Meetings in Iowa}
\format{A data frame with 1549 rows and 13 variables:
\describe{
  \item{Day}{day of the week}
  \item{Time}{time of the day}
  \item{AmPm}{morning or afternoon}
  \item{Meeting}{name of the meeting}
  \item{Location}{location}
  \item{Address}{address}
  \item{Format}{Format of the meeting - XXX this needs some additional work}
  \item{City}{city}
  \item{Type}{type of meeting}
  \item{State}{state}
  \item{lon}{geographic Longitude}
  \item{lat}{geographic Latitude}
  \item{schedule}{weekly schedule of meetings, starting with Monday at midnight.}
}}
\source{
\url{https://www.aa-iowa.org/meetings/}, \url{https://www.na-iowa.org/meetings/}
}
\usage{
meetings
}
\description{
Dataset was scraped from the AA and NA websites of Iowa by Jessie Bustin.
}
\examples{
# Location of meetings in Iowa
library(ggplot2)
library(dplyr)
meetings \%>\%
  ggplot(aes(x = lon, y = lat)) +
  geom_point(aes(colour = Type))

# Leaflet map of meetings in Iowa
library(leaflet)
library(sf)
meetings \%>\%
 leaflet() \%>\%
 addTiles() \%>\%
 addPolygons(data = ia_counties,
             weight = 1, color="#333333") \%>\%
 addCircleMarkers(meetings \%>\% filter(Type == "Narcotics Anonymous"),
                  lng = ~lon, lat = ~lat,
                  radius = 1, stroke = 0.1,
                  color = "darkorange",
                  label = ~Meeting, group="NA") \%>\%
 addCircleMarkers(meetings \%>\% filter(Type == "Alcoholics Anonymous"),
                  lng = ~lon, lat = ~lat,
                  radius = 1, stroke = 0.1,
                  color = "darkgreen",
                  label = ~Meeting, group="AA") \%>\%
 addLayersControl(overlayGroups = c("AA","NA"))
}
\keyword{datasets}
