% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parks}
\alias{parks}
\title{Location of parks in Iowa}
\format{A data frame with 1645 rows and 15 variables:
\describe{
   \item{ID}{identifier, not quite the row number}
   \item{NAME}{name of the park}
   \item{COUNTY}{name of the county}
   \item{STREET}{street name}
   \item{CITY}{name of the city}
   \item{STATE}{name of the state}
   \item{PHONE}{phone number}
   \item{ZIP}{5-digit zip code}
   \item{result_num}{1645 0s XXX delete column?}
   \item{status}{1645 OKs XXX delete column?}
   \item{formatted_}{formatted addresses - some are duplicates - XXX look into}
   \item{place_id}{identifier, based on address? - some are duplicates - XXX look into}
   \item{location_t}{categorical variable with additional details on location.}
   \item{Latitude}{geographic latitude}
   \item{Longitude}{geographic longitude}
}}
\source{
\url{https://www.mycountyparks.com/County/Default.aspx}
}
\usage{
parks
}
\description{
Dataset was scraped by Masoud Nosrati from MyCountyParks in Mar 2020, geocoding by Andrew Maloney through QGIS.
}
\examples{
# Map of parks in Iowa county using ggplot2
library(ggplot2)
library(dplyr) # for the pipe

parks \%>\%
 # filter(COUNTY == "Story county) \%>\%
  ggplot() +
    geom_point(aes(x = Longitude, y = Latitude))

# leaflet map
library(leaflet)
library(sf)

parks \%>\%
  leaflet() \%>\%
    addTiles() \%>\%
    addPolygons(data = st_transform(ia_counties, crs='+proj=longlat +datum=WGS84'),
                weight = 1, color="#333333") \%>\%
    addCircleMarkers(lng = ~Longitude, lat = ~Latitude,
                     radius = 1, stroke = 0.1,
                     label = ~NAME)
}
\keyword{datasets}
