% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{southwest_mhds}
\alias{southwest_mhds}
\title{Southwest Iowa Region Mental Health & Disability Services}
\format{
A data frame with 356 rows and 13 variables:
\describe{
  \item{category}{Service type provided at facility}
  \item{county}{Iowa county name}
  \item{related records}{Currently Unknown}
  \item{service title}{Name of the facility}
  \item{address}{Column containing location information: Street, City, State, Zip}
  \item{phone}{Phone Number for the facility}
  \item{description}{A description to help the viewer navigate which facility is best for them}
  \item{website}{Facility website link}
  \item{email}{Contact Email for facility}
  \item{expiration}{Currently Unknown}
  \item{last updated}{Date at which website information was last updated}
  \item{longitude}{geographic Longitude}
  \item{latitude}{geographic Latitude}
}
}
\source{
\url{https://www.aa-iowa.org/meetings/}
}
\usage{
southwest_mhds
}
\description{
Dataset was scraped from the MHDS website.
}
\examples{
# Location of MHDS in Iowa
library(ggplot2)
library(dplyr)
southwest_mhds \%>\%
 ggplot() +
 geom_point(aes(x = longitude, y = latitude))

# Leaflet map of meetings in Iowa
library(leaflet)
library(sf)

southwest_mhds \%>\%
 leaflet() \%>\%
 addTiles() \%>\%
 addPolygons(data = ia_counties,
             weight = 1, color="#333333") \%>\%
 addCircleMarkers(lng = ~longitude, lat = ~latitude,
                  radius = 1, stroke = 0.1,
                  label = ~southwest_mhds)
}
\keyword{datasets}
